move isnested and istestfunction, add task_name to resolver
