+++ tests/flytekit/unit/core/test_serialization.py
@@ -287,7 +287,6 @@ def test_serialization_command1():
     assert srz_t.container.args[-8:] == [
         "--resolver",
         "flytekit.core.python_auto_container.default_task_resolver",
-        "--resolver-args",
         "--",
         "task-module",
         "test_serialization",  # when unit testing, t1.task_function.__module__ just gives this file
